feat: Created a ClientGenerator class
This class was created so that we're always using one instance of both the OkHttpClient and Retrofit. More specifically, it ensures that only one networking socket is opened for all of our HTTP requests. The idea was taken from here:

https://futurestud.io/tutorials/retrofit-2-creating-a-sustainable-android-client

All variables (as well as methods) are made static so that there is only one instance of them at a time.