feat: Adjusted model objects
I noticed that I got really lucky the first time I got this to
work. This was because the app was still working... But my model
objects didn't exactly line up with the JSON response. I've
since adjusted my models by first removing the Tracks class, which
was a bit redundant. After that, I adjusted the Paging class and
created a PagingTracks subclass, which will actually store the
Track objects.